feat: day 9 solved with 2d vector and then colouring it
I got really uncomfortable with the possibility of ambiguous input where
it is not clear which basin a number belongs to. E.g. a saddle contour
can result in two possible basins. I just assumed pointing to the lowest
and if there's a tie choose any will result in the correct answer.

This exercise also got me to realize what a pain it is to instantiate
2D vectors where the dimensions are not determined at compile time. I do
wonder if there's an easier way.