feat: day 19 - brute force
This was really quite a challenge.
I first got the 24 matrices wrong, and that wasted some of my time.
Then I was fighting the compiler to access multiple elements in the same
vector. It turns out rust explicitly prevents this.

If there is a borrow on a vector, the whole vector is "locked". There
are things like split_mut - which is also quite confusing for
first-timer like me but that doesn't give an easy way to partition a
vector into two separate iterators.

In the end, that forced me to change to a different approach: keeping a
vector of solved sensors. This gives me an iterator over the solved
ones, and allows me to use the initial one for the unsolved sensors.

The algorithm here is kind of dumb: For each sensor, calculate all
rotated positions, and then all relative differences in the rotated
position. Then try to match this relative differences with the pool of
solved sensor. If there is a match of >= 12, record what the rotation
was, and the translation needed from the frame of reference chosen by
the solved sensor. Translations are additive, so it all works out
naturally.