feat: day 22: intersecting cube
Boy I found this really challenging. Intersecting segments is is really
tricky to argue about, and this was for 3d instead of 2d intersection.

I was hoping for a clever way to get around this, but the approach I
went for was break the cuboid A - cuboid B up into smaller cuboids and
then count the area for those.

A large part of my time was spent trying to figure out how to implement
that "subtraction". After I did that, I tried to run it agains the
examples and they all returned fine. Then when I ran for the main input,
it just took forever. It took me another long time before I realized the
crux that my implementation was too dumb to combine smaller cuboids
together. So there was a branching factor of 26 which blows things up.

After I added the short circuit, most iteration now has a branching
factor of 1, and the code calculates the answer in about 1s.