optimize: improve dijkstra to be able to solve part 2
The old code has too high a branching factor and way
too many clones.
The culprit comes from moving from room to hallway.
The previous code is considering iterations that are blocked.

I had expected the checking of whether it's reachable was quick enough, but interestingly it's not!.

Other optimizations:
* Instead of storing a map of state -> cost, which requires
cloning the state, I've added a sig/hash function, which makes it faster.
* I've changed dequeue to vec which is simpler and also resulted in quicker runs.