feat: day 24 - dfs and brute force
The hard part here is actually getting my hands dirty and trying to
understand what the instructions are doing and reverse engineering a
higher level algo. Once that is done, the problem reduces to a DFS
search question.

I had an uneasy feeling thinking about how to handle the general case.
i.e. what if the instructions did not have repeatable blocks. Thankfully
part 2 does not ask for that. I suppose the expected output (search
area) would be too big for the general case. It would be an abstract
machine that would need to be able to backtrack and prune aggressively
in order to cover 9^14 area.

This was quite fun - to reverse engineer the code.