feat: day 25: simulating the grid
Oh this is a nice easy one to finish. Pretty straight forward here.
Only thing to note is that I've lumped right and down as the same step
so it clones the grid twice in that step. I probably can reduce the
number of clones if I want to optimize.

I could probably also just keep track of the last column and last row
from the previous state to avoid cloning the whole grid. But this is
good enough for now.