feat: solve day 1-3 2025
Comments

Day 1: Feels like there should be a cleaner way. It was more natural to
think in terms of addition so I transform the subtraction case into
addition. It was a bit tricky to get the 0 case right. 5 -> 0 -> 95.

Day 2: I feel happy being to find the natural extension of 2 repetition
-> n repetition. It was a bit tricky to get the arithmetic right for
getting the lower and upper bounds to scan.

Day 3: Code is definitely no efficient, took about 8s to run for part 2.
I think I need to optimize the code by looking at only the first max
rather than all max (lazy to type). This is valid by a greedy argument.